Check Network Status

+++ GalleryAppMvvm/app/src/main/java/com/example/galleryappmvvm/viewmodel/FullScreenViewModel.kt
@@ -1,17 +1,24 @@
 package com.example.galleryappmvvm.viewmodel
 
+import Utils.isNetworkAvailable
+import android.content.Context
 import androidx.lifecycle.MediatorLiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.Observer
 import androidx.lifecycle.ViewModel
 import com.example.galleryappmvvm.model.Repository
 private val TAG = FullScreenViewModel::class.java.simpleName
-class FullScreenViewModel(private var repository: Repository): ViewModel(){
+class FullScreenViewModel(private var context: Context,private var repository: Repository): ViewModel(){
     private var status = MediatorLiveData<DelteStatus>()
     private var errMessage = MutableLiveData<String>()
 
     fun deleteImage(imageUrl:String, categoryId:String, currentImageId:String){
         status.value = DelteStatus.SHOW_PROGRESS
+        if(!context.isNetworkAvailable()){
+            status.value = DelteStatus.HIDE_PROGRESS
+            errMessage.value = "Network not available"
+            return
+        }
         status.addSource(repository.deleteImage(imageUrl,categoryId, currentImageId), Observer {
             it.onSuccess {
                 status.value = DelteStatus.HIDE_PROGRESS
@@ -23,13 +30,14 @@ class FullScreenViewModel(private var repository: Repository): ViewModel(){
             }
         })
     }
+
     fun getErrorMessage(): MutableLiveData<String> {
         return errMessage
     }
+
     fun getDeleteStatus(): MediatorLiveData<DelteStatus> {
         return status
     }
-
 
     enum class DelteStatus {
         SHOW_PROGRESS,
